Removing superflous contract.rs file
